+++ examples/nextjs/app/spotify-audio/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from 'next';
 import SpotifyAudio from 'spotify-audio-element/react';
+import Player from '../player';
 
 export const metadata: Metadata = {
   title: 'Spotify Audio - Media Elements',
@@ -9,11 +10,15 @@ export default function Page() {
   return (
     <>
       <section>
-        <SpotifyAudio
-          style={{ width: 500 }}
-          src="https://open.spotify.com/episode/7makk4oTQel546B0PZlDM5"
-          controls
-        ></SpotifyAudio>
+        <Player
+          as={SpotifyAudio}
+          src="https://open.spotify.com/episode/5Jo9ncrz2liWiKj8inZwD2"
+          config={{
+            startAt: 20,
+            theme: 'dark',
+            preferVideo: true,
+          }}
+        />
       </section>
     </>
   );